re-populates the Redis cache on GET (for published items)
This commit introduces a mechanism to re-populate the cache when we
encounter GET scenarios where data is **present** in Postgres, but
**absent** in Redis.

Merging this change will mean the cache is no longer stateful,
a data loss will result in a re-population of the cache without intervention.

Prior to this commit, the Redis cache was only acting as a cache for
components modified **after** the cache's inception. As a result, it was
important to have a disaster recovery plan in the case of a cache loss.