logs failed Redis caching (on PSQL GET) in a warning
We capture errors when attempting to PUT to Redis after a Postgres GET.
A failed PUT doesn't break the application, but it does suggest the
caching mechanism is broken. Logging these is preferable to silently
masking the issue. We could log at the ERROR level, but I think we
should reserve that for show-stopping issues.