uses global Redis keyspace rather than a hashmap
Utilizing a hashmap can be an effective strategy for reducing the
memory footprint of objects in long-term storage.

However... it's only efficient when the number of items is less than
the value of `hash-max-ziplist-entries`, which is 512 by default.

In NYMag's case, one hash map was storing 13MM+ items.

This configuration failed to take advantage of the memory benefits of
HMAP and completely eliminated the posibility of setting a TTL on keys
(something useful for a cache).

This commit will utilize the global keyspace for cached items by simply
removing the `h` in front of most commands.