enables expiration in Redis cache via CLAY_STORAGE_POSTGRES_CACHE_TTL
This commit enables a global TTL for cached components by setting the
`CLAY_STORAGE_POSTGRES_CACHE_TTL` environment variable.

Setting a TTL will help reduce the persistent cache size and massive
storage requirements of keeping the entire database in memory.

A TTL will also guarantee eventual consistency with the underlying
persistence layer (Postgres), even if things are modified out-of-band.

The implementation here utilizes Redis pipelines to eliminate the
overhead of the TCP handshake for each command sent. We have to do this
for efficiency because there is not `MSETEX` method in Redis (to set
multiple keys with TTLs at once).

One downside to the pipeline approach is that it may not be
well-supported when running Redis in Cluster mode, which could be a
valid use case for very large or elastic Amphora distributions.

The changes to `redis/index.test.js` are not exactly providing the
assurance we would want about how this all works. I think it
underscores a need for a new testing paradigm in this library (that
includes a `postgres` and `redis` running in Docker). There are so many
mocks present that we are often just testing whether the code we wrote is
calling the functions we wrote them to call.